remove packed-git.el
This library was written so that the tools used maintain the Emacsmirror
could extract metadata from a package without having to to checkout a
revision.  These tools have now been rewritten to work with the files in
the working tree, which makes things a lot simpler.

So I no longer have a need for this and am removing it.  I also don't
think this was a good ideas in the first place.  Eventually I will write
`git-handler.el' which will allow doing this sort of thing without
having to write any specialized variants that act on blobs instead of
files.

Meanwhile, if you really need this, you can just use the last version
at 0.4.0:package-git.el.