Remove support for ignoring certain libraries and directories
Libraries whose name matched constant `packed-ignore-library-regexp'
or which are located inside directories whose name matched constant
`packed-ignore-directory-regexp' used to be silently ignored.  This
was done for the benefit of the Emacsmirror.

Most other feature issues of mirrored packages have been dealt with by
now, either in the upstream repository or by explicitly noting it in
the Epkg database.  And it is time to no longer silently ignore the
issues that the mentioned constants helped hiding.  For the time being
they are recorded in the Epkg database as issues, eventually many of
them will hopefully be fixed upstream.

For other users of `packed-libraries' it never was a good idea to
silently ignore certain problematic files and directories, so this
commit removes this kludge by removing the constants without a
replacement and by adjusting their callers accordingly.