Remove PACKAGE argument of functions that list libraries
The PACKAGE argument was useful to avoid false-positives when deciding
which files and directories that should be silently ignored when
determining the libraries belonging to a package.  Support for that
has been removed in the previous commit, so passing the package name
is mostly useless now.

After the previous commit, one place remained where the package name
was still consulted: When determining whether a file is the package's
"Elpa multi-file package description file".  Here it was useful to
match against "PACKAGE-pkg\\.el$" instead of just "-pkg\\.el$" to
avoid false-positives.

However avoiding "-pkg.el" files was unnecessary altogether because
these files don't provide a feature and therefore are not considered
libraries anyway.  So this additional filter can be dropped, and now
passing PACKAGE around would be completely useless.